Created StateComponent and moved code from main to the new file PlayerContorlSystem.cpp
The StateComponent is for further changes to character control and actions. MovementComponent can be discarded since it's using body from box2d as well, but it could be useful if structure changes are happening to physicsComponent.

Project Property changes: added GLFW_INCLUDE_NONE to project's preprocessor directives.This will stop GLFW from including the OpenGL headers, which was causing an error.